# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Elite Starx earning platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /ads/watch:
    post:
      operationId: watchAd
      tags: [ads]
      summary: Record ad watch and earn
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WatchAdBody"
      responses:
        "200":
          description: Earning recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WatchAdResponse"
        "400":
          description: Daily limit reached or invalid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /plans/buy:
    post:
      operationId: buyPlan
      tags: [users]
      summary: Purchase the active plan
      responses:
        "200":
          description: Plan purchased successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          description: Insufficient balance or plan already active
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /rewards/claim:
    post:
      operationId: claimReward
      tags: [rewards]
      summary: Claim a milestone reward
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                referrals:
                  type: integer
              required:
                - referrals
      responses:
        "200":
          description: Reward claimed successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          description: Invalid request or already claimed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
    RegisterBody:
      type: object
      properties:
        username:
          type: string
        email:
          type: string
        password:
          type: string
        referralCode:
          type: ["string", "null"]
        whatsappNumber:
          type: string
      required:
        - username
        - email
        - password
        - whatsappNumber
    LoginBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
      required:
        - user
        - token
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
        referralCode:
          type: string
        uplinerName:
          type: ["string", "null"]
        balance:
          type: number
        totalEarnings:
          type: number
        totalWithdrawn:
          type: number
        referralCount:
          type: integer
        isAdmin:
          type: boolean
        isBlocked:
          type: boolean
        hasActivePlan:
          type: boolean
        whatsappNumber:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - username
        - email
        - referralCode
        - balance
        - totalEarnings
        - totalWithdrawn
        - referralCount
        - isAdmin
        - isBlocked
        - hasActivePlan
        - createdAt
...
    WatchAdBody:
      type: object
      properties:
        adId:
          type: string
        watchDuration:
          type: integer
      required:
        - adId
        - watchDuration
    WatchAdResponse:
      type: object
      properties:
        earned:
          type: number
        totalToday:
          type: number
        adsWatchedToday:
          type: integer
        dailyLimitReached:
          type: boolean
        newBalance:
          type: number
      required:
        - earned
        - totalToday
        - adsWatchedToday
        - dailyLimitReached
        - newBalance
...