# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.8.0 🍺
 * Do not edit manually.
 * Api
 * Elite Starx earning platform API
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  id: number;
  username: string;
  email: string;
  referralCode: string;
  /** @nullable */
  uplinerName?: string | null;
  balance: number;
  totalEarnings: number;
  totalWithdrawn: number;
  referralCount: number;
  isAdmin: boolean;
  isBlocked: boolean;
  hasActivePlan: boolean;
  /** @nullable */
  whatsappNumber?: string | null;
  createdAt: string;
}